Refactor and restructure LaraFiles package.
Replaced `LaraFilesProvider` with `LaraFileServiceProvider` and introduced an observer for the `LaraFile` model. Moved files into more logical namespaces and directory structures (e.g., `Models` and `Observers`). Updated migration file locations and adjusted file references throughout the package. Added new `.gitignore` entries for `composer.lock` and `phpunit.xml`.
