--- src/LaraFilesProvider.php
@@ -1,32 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DjurovicIgoor\LaraFiles;
-
-use Illuminate\Support\ServiceProvider;
-
-class LaraFilesProvider extends ServiceProvider
-{
-    /**
-     * Bootstrap the application services.
-     */
-    public function boot(): void
-    {
-        $this->publishes([
-            __DIR__.'/../config/lara-files.php' => config_path('lara-files.php'),
-        ], 'config');
-
-        $this->publishes([
-            __DIR__.'/database/migrations/' => base_path('/database/migrations'),
-        ], 'migrations');
-    }
-
-    /**
-     * Register the application services.
-     */
-    public function register(): void
-    {
-        $this->mergeConfigFrom(__DIR__.'/../config/lara-files.php', 'lara-files');
-    }
-}
