+++ src/Models/Observers/LaraFileObserver.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace DjurovicIgoor\LaraFiles\Models\Observers;
+
+use DjurovicIgoor\LaraFiles\Models\LaraFile;
+use Illuminate\Support\Str;
+
+class LaraFileObserver
+{
+    public function creating(LaraFile $laraFile): void
+    {
+        $laraFile->id = Str::uuid()->toString();
+    }
+}
